refactor: streamline site internals and repo artifact topology
Split the static-site builder into role-focused helper modules, move repo-local generated artifacts behind an explicit artifacts root, and update demo/docs planning surfaces to match the new topology.
