+++ devtools/mutmut_campaign.py
@@ -41,10 +41,10 @@ DEFAULT_IGNORE_PATTERNS = shutil.ignore_patterns(
     "mutants",
     "__pycache__",
     ".claude",
+    "artifacts",
     "qa_archive",
     "qa_outputs",
     "qa_2026-03-10",
-    "demos/output",
 )
